Rename modal component
